fix bug showing empty menu in case menubind.smn can't be found
LoadTVMenuAndStatus() and statusbar is called twice, first from
GetTVStatusLine(), later from GetTVMenu(), but even if it
fails early due to the inability to fopen(), InitAddSubmenu() is
called which creates an empty 'firstMenu'.
GetTVMenu then skips creating the menu altogether, as it only
does it if firstMenu isn't NULL.
firstMenu is then used as the menu, leaving it effectively blank,
with no way of exiting the editor.